:recycle: Orchestrator + refactoring
* Implement an orchestrator to build PurchaseRecap using CSV
* Call node API using object built by the CSV file
* Remove unused classes and methods

+++ backend-java/src/main/java/com/codeworks/onboarding/infrastructure/purchase/items/PurchaseItemsEntity.java
@@ -31,5 +31,5 @@ public class PurchaseItemsEntity {
     private Long buyerId;
 
     @Getter
-    private Integer unitPrice;
+    private Double unitPrice;
 }
